Fix a deadlock when an ingestion is broken.
The function that does the ingestion quit, since it makes no sense to
keep ingesting layers that are not going to be used, however, the
function providing the layers keep trying to push them into a channel
that now, nobody is reading from. Hence there are goroutines around
keeping waiting.

This could be detected by the deadlock detector mechanism in the go
runtime, however we do have a goroutine that keep work, which put off
the detector. That goroutine is the one that print that the software
keep running.